+++ packages/core/src/Plugin.ts
@@ -3,11 +3,12 @@
  * @license     https://opensource.org/licenses/MIT
  */
 
+import optimal, { Builder, Blueprint } from 'optimal';
 import Tool from './Tool';
 
 export const DEFAULT_PLUGIN_PRIORITY: number = 100;
 
-export default class Plugin<Options = {}> {
+export default class Plugin<Options extends object = {}> {
   moduleName: string = '';
 
   name: string = '';
@@ -20,9 +21,22 @@ export default class Plugin<Options = {}> {
   tool: Tool<any>;
 
   constructor(options: Partial<Options> = {}) {
-    // @ts-ignore Handle in sub-classes or bootstrap()
-    this.options = { ...options };
+    this.options = optimal(options, this.blueprint(), {
+      name: this.constructor.name,
+      unknown: true,
+    });
   }
 
+  /**
+   * Define an optimal blueprint in which to validate and build the
+   * options passed to the constructor.
+   */
+  blueprint(): { [K in keyof Options]: Builder<any> | Blueprint } {
+    return {} as any;
+  }
+
+  /**
+   * Called once the plugin has been loaded by the tool.
+   */
   bootstrap() {}
 }