+++ packages/core/tests/fixtures/plugin-exported-definition/index.js
@@ -1,5 +1,14 @@
 const path = require('path');
+const { string, number, bool } = require('optimal');
 
 const Plugin = require(path.join(__dirname, '../../../src/Plugin')).default;
 
-module.exports = class TestPlugin extends Plugin {};
+module.exports = class TestPlugin extends Plugin {
+  blueprint() {
+    return {
+      foo: string().empty(),
+      bar: bool(),
+      baz: number(),
+    };
+  }
+};