+++ packages/core/tests/fixtures/reporter/index.js
@@ -1,5 +1,12 @@
 const path = require('path');
+const { number } = require('optimal');
 
 const Reporter = require(path.join(__dirname, '../../../src/Reporter')).default;
 
-module.exports = class TestReporter extends Reporter {};
+module.exports = class TestReporter extends Reporter {
+  blueprint() {
+    return {
+      fps: number(),
+    };
+  }
+};