+++ packages/core/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Added `Tool#isCI` method to check if in a CI environment.
 - Added `Tool#isPluginEnabled` method to check if a plugin by type has been enabled.
+- Added `Plugin#blueprint`, `Reporter#blueprint`, and `Routine#blueprint` as a means to validate and
+  build options passed to the constructor.
 - Registered plugins can now define a list of custom NPM `scopes` to lookup modules in.
 
 #### 🛠 Internal